screen-orientation/browsing-contexts.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<script>
  async function makeIFrame(src = "./resources/blank.html") {
    const iframe = document.createElement("iframe");
    document.body.appendChild(iframe);
    await new Promise(r => {
      if (iframe.contentDocument.readyState === "complete") {
        return r(); 
      }
      iframe.onload = r;
    });
    return iframe;
  }

  promise_test(async t => {
    const iframe = await makeIFrame();
    await test_driver.bless("request full screen", () => {
      return document.documentElement.requestFullscreen();
    });
    const { type: currentOrientation } = screen.orientation;
    const isPortrait = currentOrientation.includes("portrait");
    const newOrientation = `${isPortrait ? "landscape" : "portrait"}`;
    const promiseToLockIFrame = iframe.contentWindow.screen.orientation.lock(
      `${newOrientation}-secondary`
    );
    const p = screen.orientation.lock(`${newOrientation}-primary`);
    await promise_rejects(
      t,
      "AbortError",
      promiseToLockIFrame,
      "Descendant browsing context's pending promise is rejected"
    );
    await p;
    screen.orientation.unlock();
    await document.exitFullscreen();
    iframe.remove();
  }, "Browsing contexts test");

  promise_test(async t => {
    const iframe = await makeIFrame();
    const promiseToChange = new Promise(resolve => {
      screen.orientation.addEventListener("change", resolve);
    });
    await test_driver.bless("request full screen", () => {
      return document.documentElement.requestFullscreen();
    });
    const { type: originalOrientation } = screen.orientation;
    const isPortrait = originalOrientation.startsWith("portrait");
    const newOrientation = `${isPortrait ? "landscape" : "portrait"}-primary`;
    await iframe.contentWindow.screen.orientation.lock(newOrientation);
    window.location.href = "https://not-web-platform.test:8443";
    await promiseToChange;
    assert_equals(
      screen.orientation.type,
      originalOrientation,
      "Navigating from iframe to new top-level browsing context unlocks orientation"
    );
    screen.orientation.unlock();
    await document.exitFullscreen();
    iframe.remove();
  }, "Navigating a nested browsing context causes the orientation to unlock()");

  promise_test(async t => {
    const promiseToChange = new Promise(resolve => {
      screen.orientation.addEventListener("change", resolve);
    });
    await test_driver.bless("request full screen", () => {
      return document.documentElement.requestFullscreen();
    });
    const { type: originalOrientation } = screen.orientation;
    const isPortrait = originalOrientation.startsWith("portrait");
    const newOrientation = `${isPortrait ? "landscape" : "portrait"}-primary`;
    const popWindow = window.open("./resources/blank.html","popup test","width=500,height=300");
    popWindow.onload = function navigation() {
      popWindow.location.href = "https://not-web-platform.test:8443"
    }
    await promiseToChange;
    assert_equals(
      screen.orientation.type,
      originalOrientation,
      "Opening pop-up screen unlocks orientation"
    );
    screen.orientation.unlock();
    window.close();
    await document.exitFullscreen();
  }, "Opening pop-up window test");
</script>
